pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hitchhiker"
version = "0.0.1"
description = ""
requires-python = ">=3.7"
dependencies = [
    "click>=8,<9",
    "tomlkit>=0.12.0,<1",
    "dotty-dict>=1.3.1,<2",
    "PyGithub>=1.59.1,<2",
]

[project.scripts]
hitchhiker = "hitchhiker.cli:cli.cli"

[project.optional-dependencies]
release = [
    "gitpython>=3.1.32,<4",
]
odoo = [
    "click-odoo>=1.6,<2",
]
test = [
    "pytest>=7,<8",
]
copier = [
    "copier>=8.3.0,<9",
]

[tool.pytest.ini_options]
python_files = ["tests/*.py", "tests/**/*.py"]

[tool.pylint.'MESSAGES CONTROL']
disable = "C0114, C0301"

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
major_on_zero = false

[tool.semantic_release.remote]
token = { env = "GH_TOKEN" }

[tool.semantic_release.branches.main]
match = "main"
prerelease = false
